backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3
pandas==2.2.2
python-dotenv==1.0.1
